pyproject.toml
[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[project]
name = "moonraker-owl"
version = "0.1.0-dev"
description = "Owl Cloud companion agent for Moonraker"
readme = "README.md"
license = { text = "MIT" }
authors = [
  { name = "Owl Engineering" }
]
requires-python = ">=3.10"
dependencies = [
  "aiohttp>=3.9",
  "paho-mqtt>=1.6",
  "backoff>=2.2",
  "pydantic>=2.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-asyncio>=0.23",
  "pytest-aiohttp>=1.0",
  "ruff>=0.6",
]

[project.scripts]
moonraker-owl = "moonraker_owl.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["moonraker_owl"]
